# openapi/flora.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "flora API",
    "description": "Manages @Flora guild deployments, User Bots, Server Custom Bots, authentication, builds, secrets, key-value stores, logs, and operator orchestration.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "flora API base path"
    }
  ],
  "paths": {
...
    "/custom-bots/": {
      "get": {
        "tags": ["User Bots"],
        "summary": "List User Bots",
        "description": "Returns every User Bot owned by the authenticated user, including deployment and gateway status. The User Bots feature must be enabled for the account.",
        "operationId": "list_custom_bots",
        "responses": {
          "200": {
            "description": "User Bots owned by the authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomBotResponse"
                  }
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["User Bots"],
        "summary": "Create a User Bot",
        "description": "Validates a Discord bot token, encrypts it at rest, and creates a User Bot owned by the authenticated user. The plaintext token is never returned. A deployment must be uploaded separately before the gateway starts.",
        "operationId": "create_custom_bot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomBotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User Bot created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomBotResponse"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/custom-bots/{bot_id}": {
      "get": {
        "tags": ["User Bots"],
        "summary": "Get a User Bot",
        "description": "Returns metadata and runtime status for a User Bot owned by the authenticated user. Tokens and other encrypted credentials are never included.",
        "operationId": "get_custom_bot",
        "parameters": [
          {
            "name": "bot_id",
            "in": "path",
            "description": "User Bot ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Bot metadata and status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomBotResponse"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["User Bots"],
        "summary": "Delete a User Bot",
        "description": "Stops the User Bot gateway, removes its active deployment, and deletes its encrypted token metadata. The operation is restricted to the owning user.",
        "operationId": "delete_custom_bot",
        "parameters": [
          {
            "name": "bot_id",
            "in": "path",
            "description": "User Bot ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Bot deleted"
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/custom-bots/{bot_id}/deployment": {
      "get": {
        "tags": ["User Bots"],
        "summary": "Get a User Bot deployment",
        "description": "Returns the active deployment source and bundle for a User Bot owned by the authenticated user. Returns a not-found problem when the bot has not been deployed.",
        "operationId": "get_custom_bot_deployment",
        "parameters": [
          {
            "name": "bot_id",
            "in": "path",
            "description": "User Bot ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active User Bot deployment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomBotDeploymentResponse"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["User Bots"],
        "summary": "Deploy a User Bot",
        "description": "Builds or accepts the supplied bundle, atomically replaces the User Bot isolate, records a deployment revision, and starts its Discord gateway. If deployment fails, the previous runtime and persisted deployment are restored.",
        "operationId": "deploy_custom_bot",
        "parameters": [
          {
            "name": "bot_id",
            "in": "path",
            "description": "User Bot ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Active User Bot deployment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomBotDeploymentResponse"
                }
              }
            }
          },
          "default": {
            "description": "The request failed. Inspect the HTTP status and Problem Details response for recovery guidance.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCustomBotRequest": {
        "type": "object",
        "required": ["token"],
        "properties": {
          "label": {
            "type": ["string", "null"]
          },
          "token": {
            "type": "string",
            "description": "Discord bot token. The token is validated, encrypted, and never returned."
          }
        }
      },
...
      "CustomBotDeploymentResponse": {
        "type": "object",
        "required": ["entry", "created_at", "updated_at"],
        "properties": {
          "bundle": {
            "type": ["string", "null"]
          },
          "created_at": {
            "type": "string"
          },
          "entry": {
            "type": "string"
          },
          "files": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/DeploymentFile"
            }
          },
          "source_map": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeploymentSourceMapFile"
              }
            ]
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "CustomBotResponse": {
        "type": "object",
        "required": [
          "id",
          "bot_user_id",
          "bot_username",
          "application_id",
          "has_deployment",
          "running",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "application_id": {
            "type": "string"
          },
          "bot_user_id": {
            "type": "string"
          },
          "bot_username": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "has_deployment": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "label": {
            "type": ["string", "null"]
          },
          "running": {
            "type": "boolean"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "DeploymentFile": {
        "type": "object",
        "required": ["path", "contents"],
        "properties": {
          "contents": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "DeploymentRequest": {
        "type": "object",
        "required": ["entry"],
        "properties": {
          "build_id": {
            "type": ["string", "null"]
          },
          "bundle": {
            "type": ["string", "null"]
          },
          "entry": {
            "type": "string"
          },
          "files": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/DeploymentFile"
            }
          },
          "source_map": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeploymentSourceMapFile"
              }
            ]
          }
        }
      },
...
      "DeploymentSourceMapFile": {
        "type": "object",
        "required": ["path", "contents"],
        "properties": {
          "contents": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status", "detail", "instance", "request_id"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Actionable explanation of this specific failure occurrence."
          },
          "instance": {
            "type": "string",
            "description": "Request path identifying where this failure occurred."
          },
          "request_id": {
            "type": "string",
            "description": "Correlation identifier also returned in the `X-Request-ID` header."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code returned for this failure occurrence.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Stable, human-readable summary of the failure category."
          },
          "type": {
            "type": "string",
            "description": "Stable relative URI identifying the category of failure."
          },
          "violations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemViolation"
            },
            "description": "Field-level validation failures, when applicable."
          }
        }
      },
      "ProblemViolation": {
        "type": "object",
        "required": ["field", "message"],
        "properties": {
          "field": {
            "type": "string",
            "description": "Identifies the invalid request field using its JSON or parameter path."
          },
          "message": {
            "type": "string",
            "description": "Explains how to correct the invalid field."
          }
        }
      },
...
    }
  },
...
}